components: Promote text
