+++ packages/components/src/ui/__storybook-utils/example-grid.js
@@ -3,7 +3,7 @@
  */
 import { Grid } from '../grid';
 import { Surface } from '../surface';
-import { Text } from '../text';
+import Text from '../../text';
 import { View } from '../view';
 import { VStack } from '../v-stack';
