+++ packages/components/src/ui/card/stories/index.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { Grid } from '../../grid';
-import { Text } from '../../text';
+import Text from '../../../text';
 import { VStack } from '../../v-stack';
 import { Card, CardBody, CardFooter } from '../index';
 import { Divider } from '../../divider';
