--- packages/components/src/ui/text/index.js
@@ -1,2 +1,0 @@
-export { default as Text } from './component';
-export { default as useText } from './hook';
